Extract functionalities from storage_test and postsql_test to storage_handler
