+++ cmd/broker/storage_test.go
@@ -1,28 +1,13 @@
 package main
 
 import (
-	"fmt"
 	"log"
 	"os"
-	"strings"
 	"testing"
 	"time"
 
-	"github.com/gocraft/dbr"
 	"github.com/kyma-project/kyma-environment-broker/internal"
-	"github.com/kyma-project/kyma-environment-broker/internal/events"
 	"github.com/kyma-project/kyma-environment-broker/internal/storage"
-	"github.com/kyma-project/kyma-environment-broker/internal/storage/postsql"
-	"github.com/sirupsen/logrus"
-	"golang.org/x/exp/slices"
-)
-
-type migrationOrder int
-
-const (
-	migrationsPath                = "./../../resources/keb/migrations/"
-	Up             migrationOrder = iota
-	Down
 )
 
 func brokerStorageTestConfig() storage.Config {
@@ -85,133 +70,9 @@ func TestMain(m *testing.M) {
 	exitVal = m.Run()
 }
 
-func GetStorageForTests() (func() error, storage.BrokerStorage, error) {
+func GetStorageForE2ETests() (func() error, storage.BrokerStorage, error) {
 	if os.Getenv("DB_IN_MEMORY") == "true" {
 		return nil, storage.NewMemoryStorage(), nil
 	}
-	config := brokerStorageTestConfig()
-	storage, connection, err := storage.NewFromConfig(config, events.Config{}, storage.NewEncrypter(config.SecretKey), logrus.StandardLogger())
-	if err != nil {
-		return nil, nil, fmt.Errorf("while creating storage: %w", err)
-	}
-	if connection == nil {
-		return nil, nil, fmt.Errorf("connection is nil")
-	}
-	if storage == nil {
-		return nil, nil, fmt.Errorf("storage is nil")
-	}
-
-	failOnIncorrectDB(connection, config)
-	failOnNotEmptyDb(connection)
-
-	err = runMigrations(connection, Up)
-	if err != nil {
-		return nil, nil, fmt.Errorf("while applying migration files: %w", err)
-	}
-	log.Println("db created")
-
-	cleanup := func() error {
-		failOnIncorrectDB(connection, config)
-		fmt.Println("cleaning up")
-		err := runMigrations(connection, Down)
-		if err != nil {
-			return fmt.Errorf("failed to clear DB tables: %w", err)
-		}
-		fmt.Println("cleaned up")
-		return nil
-	}
-
-	return cleanup, storage, nil
-}
-
-func runMigrations(connection *dbr.Connection, order migrationOrder) error {
-	if order != Up && order != Down {
-		return fmt.Errorf("unknown migration order")
-	}
-
-	files, err := os.ReadDir(migrationsPath)
-	if err != nil {
-		return fmt.Errorf("while reading migration data: %w in directory :%s", err, migrationsPath)
-	}
-
-	suffix := ""
-	if order == Down {
-		suffix = "down.sql"
-		slices.Reverse(files)
-	}
-
-	if order == Up {
-		suffix = "up.sql"
-	}
-
-	for _, file := range files {
-		if strings.HasSuffix(file.Name(), suffix) {
-			content, err := os.ReadFile(migrationsPath + file.Name())
-			if err != nil {
-				return fmt.Errorf("while reading migration files: %w file: %s", err, file.Name())
-			}
-			if _, err = connection.Exec(string(content)); err != nil {
-				return fmt.Errorf("while applying migration files: %w file: %s", err, file.Name())
-			}
-		}
-	}
-
-	return nil
-}
-
-func failOnIncorrectDB(db *dbr.Connection, config storage.Config) {
-	if db == nil {
-		panic("db is nil")
-	}
-	row := db.QueryRow("SELECT CURRENT_DATABASE();")
-	var result string
-	err := row.Scan(&result)
-	if err != nil || result == "" {
-		panic(fmt.Sprintf("cannot check if db has test prefix. %s", err.Error()))
-	}
-	has := strings.HasPrefix(result, "test")
-	if !has {
-		panic("db has not test prefix")
-	}
-	equal := strings.EqualFold(result, config.Name)
-	if !equal {
-		panic(fmt.Sprintf("db: %s is not equal to config: %s", result, config.Name))
-	}
-}
-
-func failOnNotEmptyDb(db *dbr.Connection) {
-	tableExists := func(table string) (bool, error) {
-		var exists bool
-		result := db.QueryRow(fmt.Sprintf(
-			`
-			SELECT EXISTS (
-			    SELECT FROM
-			        pg_tables
-			    WHERE
-			        schemaname = 'public' AND
-			        tablename  = '%s'
-			);`,
-			table))
-		err := result.Scan(&exists)
-		if err != nil {
-			return false, err
-		}
-		return exists, nil
-	}
-
-	exists, err := tableExists(postsql.OperationTableName)
-	if err != nil {
-		panic(fmt.Sprintf("cannot verify if table %s exists: %s", postsql.OperationTableName, err.Error()))
-	}
-	if exists {
-		panic(fmt.Sprintf("in db, data for %s are in table", postsql.OperationTableName))
-	}
-
-	exists, err = tableExists(postsql.InstancesTableName)
-	if err != nil {
-		panic(fmt.Sprintf("cannot verify if table %s exists: %s", postsql.InstancesTableName, err.Error()))
-	}
-	if exists {
-		panic(fmt.Sprintf("in db, data for %s are in table", postsql.InstancesTableName))
-	}
+	return storage.GetStorageForTest(brokerStorageTestConfig())
 }
